Payment App Update
Card Module is DONE.
All the functions that was required work as intended now, ALTHOUGH getCardPAN must take a no-space PAN to work correctly. I will leave it like this for now and try to optimize it using the not-so-completed auxiliary function noSpacesPAN. But that can be ignored for now until everything works as intended in the project.